refactor: extract PrivacyDataExportManifest.swift from PrivacyDataExportTypes.swift
Splits the `PrivacyDataExportManifest` Encodable header out of
PrivacyDataExportTypes.swift so that file holds only the snapshot
value types.

Byte-identical move of lines 3-9. Both files import Foundation.

PrivacyDataExportTypes.swift: 81 → 73 lines. Tests: 667.
