+++ Sources/BugNarrator/Services/PrivacyDataExportManifest.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+struct PrivacyDataExportManifest: Encodable {
+    let generatedAt: Date
+    let sessionCount: Int
+    let includesSecrets: Bool
+    let exportedFiles: [String]
+    let notes: [String]
+}
