+++ Sources/BugNarrator/Services/PrivacyDataExportTypes.swift
@@ -1,13 +1,5 @@
 import Foundation
 
-struct PrivacyDataExportManifest: Encodable {
-    let generatedAt: Date
-    let sessionCount: Int
-    let includesSecrets: Bool
-    let exportedFiles: [String]
-    let notes: [String]
-}
-
 struct PrivacyDataExportSettingsSnapshot: Codable, Equatable {
     let openAIBaseURL: String
     let transcriptionModel: String
